chore: consolidate identical domain types into @celsius/shared
The order and loyalty apps each defined the same 11 entity types
(Brand, Outlet, Member, PointTransaction, Redemption, Campaign,
SMSLog, StaffUser, TopSpender, DashboardStats, IssuedReward). Move
them to a new shared/domain.ts as the single source of truth; both
apps now re-export from @celsius/shared via their existing "@/types"
surface, so no consumer imports change.

MemberBrand and Reward are deliberately left app-local: loyalty's are
supersets (tier fields; combo/override reward mechanics and a wider
discount_type union) that the order/pickup surface doesn't support.
Merging them would wrongly widen the order types.

Type-only change (erased at build); verified order + loyalty tsc clean.
